+++ backend/src/pytorch/InterpolateArchs/RIFE/rife46IFNET.py
@@ -196,5 +196,5 @@ class IFNet(nn.Module):
 
         temp = torch.sigmoid(latest_mask)
         frame = warped_img0 * temp + warped_img1 * (1 - temp)
-        frame = frame[:, :, : self.height, : self.width]
+        frame = frame.float()[:, :, : self.height, : self.width]
         return frame
